22.08.06 업데이트

+++ [BOJ] 백준 풀이/단계별로 풀어보기 (step by step)/#15 백트래킹/14888_연산자 끼워넣기 (22.08.06 성공).py
@@ -25,7 +25,7 @@ def dfs(cur, res):
         elif i == 2:
             dfs(cur + 1, res * arr[cur])
         elif i == 3:
-            dfs(cur + 1, res / arr[cur])
+            dfs(cur + 1, int(res / arr[cur]))
         
         operators[i] += 1
